MODUL-916 - Separated toast service plugin from toast service:
