+++ src/components/form/form.plugin.ts
@@ -2,7 +2,7 @@ import { PluginObject } from 'vue';
 import { FORM_FIELD_NAME } from '../../directives/directive-names';
 import { FormClearToastBehavior, FormErrorFocusBehavior, FormErrorToastBehavior, MFormListener, MFormService } from '../../utils/form/form-service/form-service';
 import ScrollToPlugin from '../../utils/scroll-to/scroll-to';
-import ToastServicePlugin from '../../utils/toast/toast-service';
+import ToastServicePlugin from '../../utils/toast/toast-service.plugin';
 import { FORM } from '../component-names';
 import I18nPlugin from '../i18n/i18n';
 import MessagePlugin from '../message/message';
